Renamed package cli to options
